Allow path to index file to be specified.
This change permits the path to the index file to be optionally
specified.  This is useful for indexing BAM files resident in a Google
Cloud bucket.  While htslib can read a BAM from a GCS stream, bri
currently demands that the bri index be at the same path.  This change
decouples the two, allowing the index to be read from or written to
local disk while the BAM resides in the cloud.