New get, "show", and verbosity features.
Allow multiple reads be fetched from a single invocation of the get
subcommand.

Added new subcommand "show", which lists the contents of an index.  This
is useful when formulating a scatter-gather plan to process pieces of a
BAM file in parallel, and is much faster than fetching the equivalent
information from the BAM file itself.

Added a verbosity ("-v") flag to the index subcommand, enabling progress
monitoring for long-running index commands.